Consolidate classes into the same package and update the ReadMe.
