+++ src/main/java/org/openlmis/example/web/BarValidation.java
@@ -1,10 +1,9 @@
-package org.openlmis.example.annotation;
+package org.openlmis.example.web;
 
-import org.openlmis.example.validator.BarValidator;
+import org.openlmis.example.web.BarValidator;
 
 import javax.validation.Constraint;
 import javax.validation.Payload;
-import java.lang.annotation.Documented;
 import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
 
@@ -12,7 +11,7 @@ import static java.lang.annotation.ElementType.TYPE;
 import static java.lang.annotation.RetentionPolicy.RUNTIME;
 
 /*
-    Create a custom @BarValidation that may be applied to Types.
+    Create a custom @BarValidation annotation which may be applied to Types.
  */
 @Target({TYPE})
 @Retention(RUNTIME)
