+++ src/main/java/org/openlmis/example/web/NotificationController.java
@@ -1,22 +1,11 @@
 package org.openlmis.example.web;
 
-import org.openlmis.example.domain.Bar;
 import org.openlmis.example.domain.Notification;
-import org.openlmis.example.repository.ReadOnlyFooRepository;
-import org.openlmis.example.validator.NotificationValidator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.data.rest.webmvc.PersistentEntityResource;
-import org.springframework.data.rest.webmvc.PersistentEntityResourceAssembler;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.bind.support.SessionStatus;
-
-import javax.validation.ConstraintViolation;
-import javax.validation.Validation;
-import javax.validation.Validator;
-import javax.validation.ValidatorFactory;
-import java.util.Set;
 
 
 @RestController
